Saves real game, doesn't work
Takes multiple hours to run (although I did optimize it in a couple of
places) but it doesn't work.

Broke down and looked up some other techniques online and I can do this
in a MUCH simpler way. Going to implement that now.
