--- 2015/22/game-state/index.js
@@ -1,13 +1,0 @@
-const { createStore } = require('redux');
-const player = require('./player');
-const boss = require('./player');
-const overall = require('./overall'); // `round`, `turn`, and `status`
-const effects = require('./effects');
-
-const initialState = require('./initial-state');
-
-const rootReducer = combineReducers({
-
-});
-
-module.exports = createStore(rootReducer, initialState);
