--- 2015/22/input.js
@@ -1,6 +1,0 @@
-module.exports = {
-    BOSS_STATS: {
-        hp: 51,
-        damage: 9,
-    },
-};
