--- 2015/22/jsnetworkx.org/api/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>JSNetworkX API Reference</title>
-    <link rel="stylesheet" href="bootstrap.min.css">
-    <link rel="stylesheet" href="solarized_light.min.css">
-    <link rel="stylesheet" href="katex.min.css">
-    <link rel="stylesheet" href="style.css">
-    <script src="highlight.pack.js"></script>
-  </head>
-  <body>
-    <script src="bundle.js"></script>
-  </body>
-</html>
