--- 2015/22/jsnetworkx.org/api/style.css
@@ -1,198 +1,0 @@
-body {
-  padding-top: 70px;
-}
-
-.container {
-  color: #666;
-  font-size: 16px;
-}
-
-.side-index {
-  position: fixed;
-  top: 70px;
-  bottom: 0;
-  overflow-y: auto;
-  box-sizing: border-box;
-}
-
-/* Extra small devices (phones, less than 768px) */
-/* No media query since this is the default in Bootstrap */
-
-/* Small devices (tablets, 768px and up) */
-@media (min-width: 768px) {}
-
-/* Medium devices (desktops, 992px and up) */
-@media (min-width: 992px) {
-  .side-index {
-    width: 340px;
-  }
-}
-
-/* Large devices (large desktops, 1200px and up) */
-@media (min-width: 1200px) {
-  .side-index {
-    width: 300px;
-  }
-}
-
-.side-index ul {
-  list-style: none;
-  padding-left: 15px;
-}
-
-.symbol-title h2,
-.symbol-title h3,
-.symbol-title h4 {
-  margin-top: 0;
-}
-
-code {
-  background-color: transparent;
-}
-
-.symbol-title h2 code,
-.symbol-title h3 code,
-.symbol-title h4 code {
-  color: inherit;
-  padding: 0;
-}
-
-.panel-title h3,
-.panel-title h4 {
-  margin-top: 0;
-  margin-bottom: 0;
-  font-size: 16px;
-  color: inherit;
-}
-
-.symbol-parameters {
-  margin: 10px 0;
-  padding-left: 10px;
-  padding-top: 0;
-  padding-bottom: 10px;
-  border: 1px solid #DDD;
-  border-radius: 3px;
-  font-size: 14px;
-}
-
-.symbol-parameters > ul {
-  list-style: none;
-  margin-left: 10px;
-  padding-left: 0;
-}
-
-.symbol-parameters > ul > li {
-  margin-bottom: 10px;
-}
-
-.symbol-description {
-  margin: 20px 0;
-}
-
-.param-descr p {
-  margin: 0 0 0 5px;
-}
-
-.returns .param-descr p {
-  margin: 5px;
-}
-
-.class-method .symbol-description h3 {
-  font-size: 20px;
-}
-
-.class-method .symbol-description h4 {
-  font-size: 18px;
-}
-
-.param-type {
-  font-family: monospace;
-  font-style: italic;
-}
-
-.section {
-  margin-right: 20px;
-  margin-bottom: 20px;
-}
-
-.section .link {
-  position: relative;
-  padding-left: 10px;
-  margin-left: 3px;
-  max-width: 320px;
-}
-
-.section .link::before {
-  content: '\2022';
-  position: absolute;
-  left: 0px;
-}
-
-.section .summary {
-  color: #888;
-  margin-bottom: 5px;
-}
-
-.section h2 {
-  background-color: #888;
-  color: white;
-  padding: 10px;
-}
-
-.main-section > .section > .section {
-  display: block;
-}
-
-.section:not(.hasSubsections) {
-  display: inline-block;
-  vertical-align: top;
-}
-
-.class-method {
-  position: relative;
-  margin-bottom: 20px;
-}
-
-.anchor {
-  position: absolute;
-  top: -70px;
-  z-index: -1;
-}
-
-.class-method .symbol-parameters {
-  border: none;
-  border-radius: 0;
-  margin-top: 0;
-}
-
-.class-method .withDescription .symbol-parameters {
-  border-bottom: 1px solid #DDD;
-  margin: 0;
-}
-
-.class-method .symbol-parameters h4 {
-  font-size: 16px;
-  margin-top: 0;
-}
-
-.class-method .symbol-description {
-  margin: 10px 0;
-}
-
-.katex-display {
-  text-align: center;
-  display: block;
-  margin: 10px;
-}
-.katex .katex-mathml {
-  display: none;
-}
-
-.aliasInfo {
-  margin-bottom: 20px;
-  font-size: 1.2em;
-}
-
-.aliasInfo code {
-  padding: 0;
-}
