--- 2015/22/jsnetworkx.org/api/versions.json
@@ -1,3 +1,0 @@
-[
-  "v0.3.4"
-]
