--- 2015/22/part-one-bfs-2.js
@@ -1,18 +1,0 @@
-const GameGraph = require('./game-bfs-2.js');
-
-let game = new GameGraph();
-console.log('Running');
-game.buildTree();
-console.log('Number of nodes: ', game.graph.nodes().length);
-console.log('Getting shortest paths');
-let paths = game.getWinningPaths();
-let weights = paths.map(p => p.weight);
-weights.sort((a, b) => a - b);
-
-console.log(weights[0]);
-console.log('=======');
-console.log(weights);
-
-
-// console.log(game.win_nodes.length);
-// console.log(game.loss_nodes.length);
